bigger refactor wizards. Implement abstract factory class
