+++ asistente_ladm_col/gui/wizards/abs_wizard_factory.py
@@ -36,22 +36,16 @@ from ...config.general_config import (PLUGIN_NAME,
                                       LAYER)
 from ...config.help_strings import HelpStrings
 from ...config.wizards_config import WizardConfig
-from ...gui.wizards.select_features_on_map_wizard import SelectFeaturesOnMapWizard
-from ...utils.qt_utils import (enable_next_wizard,
-                               disable_next_wizard)
 from ...utils.ui import load_ui
 
 
-class MultiPageSpatialWizard(QWizard, SelectFeaturesOnMapWizard):
-
+class AbsWizardFactory(QWizard):
     set_wizard_is_open_emitted = pyqtSignal(bool)
     set_finalize_geometry_creation_enabled_emitted = pyqtSignal(bool)
 
     def __init__(self, iface, db, qgis_utils, wizard_settings):
-        self.iface = iface  # in another position breaks the constructor
-        QWizard.__init__(self)
-        SelectFeaturesOnMapWizard.__init__(self)
-
+        super(AbsWizardFactory, self).__init__()
+        self.iface = iface
         self.log = QgsApplication.messageLog()
         self._db = db
         self.qgis_utils = qgis_utils
@@ -70,101 +64,13 @@ class MultiPageSpatialWizard(QWizard, SelectFeaturesOnMapWizard):
         self.init_gui()
 
     def init_gui(self):
-        self.restore_settings()
-        self.rad_create_manually.toggled.connect(self.adjust_page_1_controls)
-        self.adjust_page_1_controls()
-
-        self.button(QWizard.NextButton).clicked.connect(self.adjust_page_2_controls)
-        self.button(QWizard.FinishButton).clicked.connect(self.finished_dialog)
-        self.button(QWizard.HelpButton).clicked.connect(self.show_help)
-        self.rejected.connect(self.close_wizard)
-        self.mMapLayerComboBox.setFilters(self.wizard_config[WizardConfig.WIZARD_MAP_LAYER_PROXY_MODEL])
+        raise NotImplementedError
 
     def adjust_page_1_controls(self):
-        self.cbo_mapping.clear()
-        self.cbo_mapping.addItem("")
-        self.cbo_mapping.addItems(self.qgis_utils.get_field_mappings_file_names(self.EDITING_LAYER_NAME))
-
-        if self.rad_refactor.isChecked():
-            self.lbl_refactor_source.setEnabled(True)
-            self.mMapLayerComboBox.setEnabled(True)
-            self.lbl_field_mapping.setEnabled(True)
-            self.cbo_mapping.setEnabled(True)
-            disable_next_wizard(self)
-            self.wizardPage1.setFinalPage(True)
-            finish_button_text = QCoreApplication.translate(self.WIZARD_NAME, "Import")
-            self.txt_help_page_1.setHtml(self.help_strings.get_refactor_help_string(self.EDITING_LAYER_NAME, True))
-            self.wizardPage1.setButtonText(QWizard.FinishButton, finish_button_text)
-        elif self.rad_create_manually.isChecked():
-            self.lbl_refactor_source.setEnabled(False)
-            self.mMapLayerComboBox.setEnabled(False)
-            self.lbl_field_mapping.setEnabled(False)
-            self.cbo_mapping.setEnabled(False)
-            self.wizardPage1.setFinalPage(False)
-            enable_next_wizard(self)
-            self.wizardPage1.setFinalPage(False)
-            finish_button_text = QCoreApplication.translate(self.WIZARD_NAME, "Create")
-            self.txt_help_page_1.setHtml(self.wizard_config[WizardConfig.WIZARD_HELP_PAGES_SETTING][WizardConfig.WIZARD_HELP_PAGE1])
-
-        self.wizardPage2.setButtonText(QWizard.FinishButton, finish_button_text)
-
-    def adjust_page_2_controls(self):
-        self.button(self.FinishButton).setDisabled(True)
-        self.txt_help_page_2.setHtml(self.wizard_config[WizardConfig.WIZARD_HELP_PAGES_SETTING][WizardConfig.WIZARD_HELP_PAGE2])
-        self.disconnect_signals()
-
-        # Load layers
-        result = self.prepare_feature_creation_layers()
-        if result is None:
-            self.close_wizard(show_message=False)
-
-        # Check if a previous features are selected
-        self.check_selected_features()
-
-        # Register select features by expression
-        if hasattr(self, 'SELECTION_BY_EXPRESSION'):
-            self.register_select_features_by_expression()
-
-        # Register select features on map
-        if hasattr(self, 'SELECTION_ON_MAP'):
-            self.register_select_feature_on_map()
-
-    def register_select_features_by_expression(self):
-        raise NotImplementedError
-
-    def register_select_feature_on_map(self):
-        raise NotImplementedError
-
-    def check_selected_features(self):
         raise NotImplementedError
 
     def finished_dialog(self):
-        self.save_settings()
-
-        if self.rad_refactor.isChecked():
-            if self.mMapLayerComboBox.currentLayer() is not None:
-                field_mapping = self.cbo_mapping.currentText()
-                res_etl_model = self.qgis_utils.show_etl_model(self._db,
-                                                               self.mMapLayerComboBox.currentLayer(),
-                                                               self.EDITING_LAYER_NAME,
-                                                               field_mapping=field_mapping)
-                if res_etl_model: # Features were added?
-                    # If the result of the etl_model is successful and we used a stored recent mapping, we delete the
-                    # previous mapping used (we give preference to the latest used mapping)
-                    if field_mapping:
-                        self.qgis_utils.delete_old_field_mapping(field_mapping)
-
-                    self.qgis_utils.save_field_mapping(self.EDITING_LAYER_NAME)
-            else:
-                self.qgis_utils.message_emitted.emit(
-                    QCoreApplication.translate(self.WIZARD_NAME,
-                                               "Select a source layer to set the field mapping to '{}'.").format(
-                        self.EDITING_LAYER_NAME),
-                    Qgis.Warning)
-
-        elif self.rad_create_manually.isChecked():
-            self.set_finalize_geometry_creation_enabled_emitted.emit(True)
-            self.prepare_feature_creation()
+        raise NotImplementedError
 
     def prepare_feature_creation(self):
         result = self.prepare_feature_creation_layers()
@@ -213,21 +119,9 @@ class MultiPageSpatialWizard(QWizard, SelectFeaturesOnMapWizard):
         return True
 
     def close_wizard(self, message=None, show_message=True):
-        if message is None:
-            message = QCoreApplication.translate(self.WIZARD_NAME, "'{}' tool has been closed.").format(self.WIZARD_TOOL_NAME)
-        if show_message:
-            self.qgis_utils.message_emitted.emit(message, Qgis.Info)
-
-        if hasattr(self, 'SELECTION_ON_MAP'):
-            self.init_map_tool()
-
-        self.set_finalize_geometry_creation_enabled_emitted.emit(False)
-        self.disconnect_signals()
-        self.set_wizard_is_open_emitted.emit(False)
-        self.close()
+        raise NotImplementedError
 
     def disconnect_signals(self):
-
         if hasattr(self, 'SELECTION_BY_EXPRESSION'):
             self.disconnect_signals_select_features_by_expression()
 
@@ -238,9 +132,6 @@ class MultiPageSpatialWizard(QWizard, SelectFeaturesOnMapWizard):
             self._layers[self.EDITING_LAYER_NAME][LAYER].committedFeaturesAdded.disconnect(self.finish_feature_creation)
         except:
             pass
-
-    def disconnect_signals_controls_select_features_on_map(self):
-        raise NotImplementedError
 
     def edit_feature(self):
         raise NotImplementedError
@@ -256,19 +147,10 @@ class MultiPageSpatialWizard(QWizard, SelectFeaturesOnMapWizard):
         raise NotImplementedError
 
     def open_form(self, layer):
-        if not layer.isEditable():
-            layer.startEditing()
-
-        self.qgis_utils.suppress_form(layer, True)
-        self.iface.actionAddFeature().trigger()
+        raise NotImplementedError
 
     def exec_form(self, layer):
-        self.set_finalize_geometry_creation_enabled_emitted.emit(False)
-
-        for id, added_feature in layer.editBuffer().addedFeatures().items():
-            feature = added_feature
-            break
-
+        feature = self.get_feature_exec_form(layer)
         dialog = self.iface.getFeatureForm(layer, feature)
         dialog.rejected.connect(self.form_rejected)
         dialog.setModal(True)
@@ -288,7 +170,10 @@ class MultiPageSpatialWizard(QWizard, SelectFeaturesOnMapWizard):
             layer.rollBack()
         self.iface.mapCanvas().refresh()
 
-    def exec_form_advance(self):
+    def get_feature_exec_form(self, layer):
+        raise NotImplementedError
+
+    def exec_form_advance(self, layer):
         raise NotImplementedError
 
     def form_rejected(self):
