+++ asistente_ladm_col/gui/wizards/multi_page_wizard_factory.py
@@ -1,0 +1,111 @@
+# -*- coding: utf-8 -*-
+"""
+/***************************************************************************
+                              Asistente LADM_COL
+                             --------------------
+        begin                : 2019-09-10
+        git sha              : :%H$
+        copyright            : (C) 2017 by Germán Carrillo
+                               (C) 2018 by Sergio Ramírez (Incige SAS)
+                               (C) 2018 by Jorge Useche (Incige SAS)
+                               (C) 2018 by Jhon Galindo
+                               (C) 2019 by Leo Cardona
+        email                : gcarrillo@linuxmail.com
+                               sergio.ramirez@incige.com
+                               naturalmentejorge@gmail.com
+                               jhonsigpjc@gmail.com
+                               leo.cardona.p@gmail.com
+ ***************************************************************************/
+/***************************************************************************
+ *                                                                         *
+ *   This program is free software; you can redistribute it and/or modify  *
+ *   it under the terms of the GNU General Public License v3.0 as          *
+ *   published by the Free Software Foundation.                            *
+ *                                                                         *
+ ***************************************************************************/
+ """
+
+from qgis.PyQt.QtCore import (QCoreApplication,
+                              pyqtSignal)
+from qgis.PyQt.QtWidgets import QWizard
+
+from .wizard_factory import WizardFactory
+from ...config.wizards_config import WizardConfig
+from ...utils.qt_utils import (enable_next_wizard,
+                               disable_next_wizard)
+
+
+class MultiPageWizardFactory(WizardFactory):
+    set_wizard_is_open_emitted = pyqtSignal(bool)
+    set_finalize_geometry_creation_enabled_emitted = pyqtSignal(bool)
+
+    def __init__(self, iface, db, qgis_utils, wizard_settings):
+        super(MultiPageWizardFactory, self).__init__(iface, db, qgis_utils, wizard_settings)
+
+    def init_gui(self):
+        self.restore_settings()
+        self.rad_create_manually.toggled.connect(self.adjust_page_1_controls)
+        self.adjust_page_1_controls()
+
+        self.button(QWizard.NextButton).clicked.connect(self.adjust_page_2_controls)
+        self.button(QWizard.FinishButton).clicked.connect(self.finished_dialog)
+        self.button(QWizard.HelpButton).clicked.connect(self.show_help)
+        self.rejected.connect(self.close_wizard)
+        self.mMapLayerComboBox.setFilters(self.wizard_config[WizardConfig.WIZARD_MAP_LAYER_PROXY_MODEL])
+
+    def adjust_page_1_controls(self):
+        self.cbo_mapping.clear()
+        self.cbo_mapping.addItem("")
+        self.cbo_mapping.addItems(self.qgis_utils.get_field_mappings_file_names(self.EDITING_LAYER_NAME))
+
+        if self.rad_refactor.isChecked():
+            self.lbl_refactor_source.setEnabled(True)
+            self.mMapLayerComboBox.setEnabled(True)
+            self.lbl_field_mapping.setEnabled(True)
+            self.cbo_mapping.setEnabled(True)
+            disable_next_wizard(self)
+            self.wizardPage1.setFinalPage(True)
+            finish_button_text = QCoreApplication.translate(self.WIZARD_NAME, "Import")
+            self.txt_help_page_1.setHtml(self.help_strings.get_refactor_help_string(self.EDITING_LAYER_NAME, False))
+            self.wizardPage1.setButtonText(QWizard.FinishButton, finish_button_text)
+        elif self.rad_create_manually.isChecked():
+            self.lbl_refactor_source.setEnabled(False)
+            self.mMapLayerComboBox.setEnabled(False)
+            self.lbl_field_mapping.setEnabled(False)
+            self.cbo_mapping.setEnabled(False)
+            enable_next_wizard(self)
+            self.wizardPage1.setFinalPage(False)
+            finish_button_text = QCoreApplication.translate(self.WIZARD_NAME, "Create")
+            self.txt_help_page_1.setHtml(self.wizard_config[WizardConfig.WIZARD_HELP_PAGES_SETTING][WizardConfig.WIZARD_HELP_PAGE1])
+
+        self.wizardPage2.setButtonText(QWizard.FinishButton, finish_button_text)
+
+    def adjust_page_2_controls(self):
+        self.button(self.FinishButton).setDisabled(True)
+        self.txt_help_page_2.setHtml(self.wizard_config[WizardConfig.WIZARD_HELP_PAGES_SETTING][WizardConfig.WIZARD_HELP_PAGE2])
+        self.disconnect_signals()
+
+        # Load layers
+        result = self.prepare_feature_creation_layers()
+        if result is None:
+            self.close_wizard(show_message=False)
+
+        # Check if a previous features are selected
+        self.check_selected_features()
+
+        # Register select features by expression
+        if hasattr(self, 'SELECTION_BY_EXPRESSION'):
+            self.register_select_features_by_expression()
+
+        # Register select features on map
+        if hasattr(self, 'SELECTION_ON_MAP'):
+            self.register_select_feature_on_map()
+
+    def advance_save(self, features):
+        raise NotImplementedError
+
+    def exec_form_advance(self, layer):
+        raise NotImplementedError
+
+    def check_selected_features(self):
+        raise NotImplementedError
