+++ asistente_ladm_col/gui/wizards/single_page_wizard_factory.py
@@ -1,0 +1,95 @@
+# -*- coding: utf-8 -*-
+"""
+/***************************************************************************
+                              Asistente LADM_COL
+                             --------------------
+        begin                : 2019-09-10
+        git sha              : :%H$
+        copyright            : (C) 2017 by Germán Carrillo
+                               (C) 2018 by Sergio Ramírez (Incige SAS)
+                               (C) 2018 by Jorge Useche (Incige SAS)
+                               (C) 2018 by Jhon Galindo
+                               (C) 2019 by Leo Cardona
+        email                : gcarrillo@linuxmail.com
+                               sergio.ramirez@incige.com
+                               naturalmentejorge@gmail.com
+                               jhonsigpjc@gmail.com
+                               leo.cardona.p@gmail.com
+ ***************************************************************************/
+/***************************************************************************
+ *                                                                         *
+ *   This program is free software; you can redistribute it and/or modify  *
+ *   it under the terms of the GNU General Public License v3.0 as          *
+ *   published by the Free Software Foundation.                            *
+ *                                                                         *
+ ***************************************************************************/
+ """
+
+from qgis.PyQt.QtCore import (QCoreApplication,
+                              pyqtSignal)
+from qgis.PyQt.QtWidgets import QWizard
+from qgis.core import Qgis
+
+from .wizard_factory import WizardFactory
+from ...config.general_config import (PLUGIN_NAME,
+                                      LAYER)
+from ...config.table_mapping_config import ID_FIELD
+from ...config.wizards_config import WizardConfig
+
+
+class SinglePageWizardFactory(WizardFactory):
+    set_wizard_is_open_emitted = pyqtSignal(bool)
+    set_finalize_geometry_creation_enabled_emitted = pyqtSignal(bool)
+
+    def __init__(self, iface, db, qgis_utils, wizard_settings):
+        super(SinglePageWizardFactory, self).__init__(iface, db, qgis_utils, wizard_settings)
+
+    def init_gui(self):
+        self.restore_settings()
+        self.rad_create_manually.toggled.connect(self.adjust_page_1_controls)
+        self.adjust_page_1_controls()
+
+        self.button(QWizard.FinishButton).clicked.connect(self.finished_dialog)
+        self.button(QWizard.HelpButton).clicked.connect(self.show_help)
+        self.rejected.connect(self.close_wizard)
+        self.mMapLayerComboBox.setFilters(self.wizard_config[WizardConfig.WIZARD_MAP_LAYER_PROXY_MODEL])
+
+    def adjust_page_1_controls(self):
+        self.cbo_mapping.clear()
+        self.cbo_mapping.addItem("")
+        self.cbo_mapping.addItems(self.qgis_utils.get_field_mappings_file_names(self.EDITING_LAYER_NAME))
+
+        if self.rad_refactor.isChecked():
+            self.lbl_refactor_source.setEnabled(True)
+            self.mMapLayerComboBox.setEnabled(True)
+            self.lbl_field_mapping.setEnabled(True)
+            self.cbo_mapping.setEnabled(True)
+            finish_button_text = QCoreApplication.translate(self.WIZARD_NAME, "Import")
+            self.txt_help_page_1.setHtml(self.help_strings.get_refactor_help_string(self.EDITING_LAYER_NAME, False))
+        elif self.rad_create_manually.isChecked():
+            self.lbl_refactor_source.setEnabled(False)
+            self.mMapLayerComboBox.setEnabled(False)
+            self.lbl_field_mapping.setEnabled(False)
+            self.cbo_mapping.setEnabled(False)
+            finish_button_text = QCoreApplication.translate(self.WIZARD_NAME, "Create")
+            self.txt_help_page_1.setHtml(self.wizard_config[WizardConfig.WIZARD_HELP_PAGES_SETTING][WizardConfig.WIZARD_HELP_PAGE1])
+
+        self.wizardPage1.setButtonText(QWizard.FinishButton, finish_button_text)
+
+    def advance_save(self, features):
+        message = QCoreApplication.translate(self.WIZARD_NAME,
+                                             "'{}' tool has been closed because an error occurred while trying to save the data.").format(self.WIZARD_TOOL_NAME)
+        fid = features[0].id()
+
+        if not self._layers[self.EDITING_LAYER_NAME][LAYER].getFeature(fid).isValid():
+            message = QCoreApplication.translate(self.WIZARD_NAME,
+                                                 "'{}' tool has been closed. Feature not found in layer {}... It's not posible create it.").format(self.WIZARD_TOOL_NAME, self.EDITING_LAYER_NAME)
+            self.log.logMessage("Feature not found in layer {} ...".format(self.EDITING_LAYER_NAME), PLUGIN_NAME, Qgis.Warning)
+        else:
+            feature_tid = self._layers[self.EDITING_LAYER_NAME][LAYER].getFeature(fid)[ID_FIELD]
+            message = QCoreApplication.translate(self.WIZARD_NAME,
+                                                 "The new {} (t_id={}) was successfully created ").format(self.WIZARD_FEATURE_NAME, feature_tid)
+        return message
+
+    def exec_form_advance(self, layer):
+        pass
