+++ asistente_ladm_col/gui/wizards/spatial_wizard_factory.py
@@ -1,0 +1,135 @@
+# -*- coding: utf-8 -*-
+"""
+/***************************************************************************
+                              Asistente LADM_COL
+                             --------------------
+        begin                : 2019-09-10
+        git sha              : :%H$
+        copyright            : (C) 2017 by Germán Carrillo
+                               (C) 2018 by Sergio Ramírez (Incige SAS)
+                               (C) 2018 by Jorge Useche (Incige SAS)
+                               (C) 2018 by Jhon Galindo
+                               (C) 2019 by Leo Cardona
+        email                : gcarrillo@linuxmail.com
+                               sergio.ramirez@incige.com
+                               naturalmentejorge@gmail.com
+                               jhonsigpjc@gmail.com
+                               leo.cardona.p@gmail.com
+ ***************************************************************************/
+/***************************************************************************
+ *                                                                         *
+ *   This program is free software; you can redistribute it and/or modify  *
+ *   it under the terms of the GNU General Public License v3.0 as          *
+ *   published by the Free Software Foundation.                            *
+ *                                                                         *
+ ***************************************************************************/
+ """
+from qgis.PyQt.QtCore import (QCoreApplication,
+                              pyqtSignal)
+from qgis.core import (QgsProject,
+                       Qgis)
+
+from .abs_wizard_factory import AbsWizardFactory
+from .select_features_on_map_wizard import SelectFeaturesOnMapWizard
+from ...config.general_config import LAYER
+
+
+class SpatialWizardFactory(AbsWizardFactory, SelectFeaturesOnMapWizard):
+    set_wizard_is_open_emitted = pyqtSignal(bool)
+    set_finalize_geometry_creation_enabled_emitted = pyqtSignal(bool)
+
+    def __init__(self, iface, db, qgis_utils, wizard_settings):
+        self.iface = iface
+        AbsWizardFactory.__init__(self, iface, db, qgis_utils, wizard_settings)
+        SelectFeaturesOnMapWizard.__init__(self)
+
+    def init_gui(self):
+        raise NotImplementedError
+
+    def adjust_page_1_controls(self):
+        raise NotImplementedError
+
+    def finished_dialog(self):
+        self.save_settings()
+
+        if self.rad_refactor.isChecked():
+            if self.mMapLayerComboBox.currentLayer() is not None:
+                field_mapping = self.cbo_mapping.currentText()
+                res_etl_model = self.qgis_utils.show_etl_model(self._db,
+                                                               self.mMapLayerComboBox.currentLayer(),
+                                                               self.EDITING_LAYER_NAME,
+                                                               field_mapping=field_mapping)
+                if res_etl_model: # Features were added?
+                    # If the result of the etl_model is successful and we used a stored recent mapping, we delete the
+                    # previous mapping used (we give preference to the latest used mapping)
+                    if field_mapping:
+                        self.qgis_utils.delete_old_field_mapping(field_mapping)
+
+                    self.qgis_utils.save_field_mapping(self.EDITING_LAYER_NAME)
+            else:
+                self.qgis_utils.message_emitted.emit(
+                    QCoreApplication.translate(self.WIZARD_NAME,
+                                               "Select a source layer to set the field mapping to '{}'.").format(
+                        self.EDITING_LAYER_NAME),
+                    Qgis.Warning)
+
+        elif self.rad_create_manually.isChecked():
+            self.set_finalize_geometry_creation_enabled_emitted.emit(True)
+            self.prepare_feature_creation()
+
+    def close_wizard(self, message=None, show_message=True):
+        if message is None:
+            message = QCoreApplication.translate(self.WIZARD_NAME, "'{}' tool has been closed.").format(self.WIZARD_TOOL_NAME)
+        if show_message:
+            self.qgis_utils.message_emitted.emit(message, Qgis.Info)
+
+        if hasattr(self, 'SELECTION_ON_MAP'):
+            self.init_map_tool()
+
+        self.set_finalize_geometry_creation_enabled_emitted.emit(False)
+        self.disconnect_signals()
+        self.set_wizard_is_open_emitted.emit(False)
+        self.close()
+
+    def edit_feature(self):
+        self.iface.layerTreeView().setCurrentLayer(self._layers[self.EDITING_LAYER_NAME][LAYER])
+        self._layers[self.EDITING_LAYER_NAME][LAYER].committedFeaturesAdded.connect(self.finish_feature_creation)
+
+        # Disable transactions groups
+        QgsProject.instance().setAutoTransaction(False)
+
+        # Activate snapping
+        self.qgis_utils.active_snapping_all_layers(tolerance=9)
+        self.open_form(self._layers[self.EDITING_LAYER_NAME][LAYER])
+
+        self.qgis_utils.message_emitted.emit(
+            QCoreApplication.translate(self.WIZARD_NAME,
+                                       "You can now start capturing {} digitizing on the map...").format(self.WIZARD_FEATURE_NAME), Qgis.Info)
+
+    def advance_save(self, features):
+        raise NotImplementedError
+
+    def open_form(self, layer):
+        if not layer.isEditable():
+            layer.startEditing()
+
+        self.qgis_utils.suppress_form(layer, True)
+        self.iface.actionAddFeature().trigger()
+
+    def get_feature_exec_form(self, layer):
+        self.set_finalize_geometry_creation_enabled_emitted.emit(False)
+        feature = None
+        for id, added_feature in layer.editBuffer().addedFeatures().items():
+            feature = added_feature
+            break
+
+        return feature
+
+    def exec_form_advance(self, layer):
+        raise NotImplementedError
+
+    def disconnect_signals_controls_select_features_on_map(self):
+        raise NotImplementedError
+
+    def register_select_feature_on_map(self):
+        raise NotImplementedError
