+++ asistente_ladm_col/gui/wizards/wizard_factory.py
@@ -1,0 +1,107 @@
+# -*- coding: utf-8 -*-
+"""
+/***************************************************************************
+                              Asistente LADM_COL
+                             --------------------
+        begin                : 2019-09-10
+        git sha              : :%H$
+        copyright            : (C) 2017 by Germán Carrillo
+                               (C) 2018 by Sergio Ramírez (Incige SAS)
+                               (C) 2018 by Jorge Useche (Incige SAS)
+                               (C) 2018 by Jhon Galindo
+                               (C) 2019 by Leo Cardona
+        email                : gcarrillo@linuxmail.com
+                               sergio.ramirez@incige.com
+                               naturalmentejorge@gmail.com
+                               jhonsigpjc@gmail.com
+                               leo.cardona.p@gmail.com
+ ***************************************************************************/
+/***************************************************************************
+ *                                                                         *
+ *   This program is free software; you can redistribute it and/or modify  *
+ *   it under the terms of the GNU General Public License v3.0 as          *
+ *   published by the Free Software Foundation.                            *
+ *                                                                         *
+ ***************************************************************************/
+ """
+from qgis.PyQt.QtCore import (QCoreApplication,
+                              pyqtSignal)
+from qgis.core import Qgis
+
+from .abs_wizard_factory import AbsWizardFactory
+from ...config.general_config import LAYER
+
+
+class WizardFactory(AbsWizardFactory):
+    set_wizard_is_open_emitted = pyqtSignal(bool)
+    set_finalize_geometry_creation_enabled_emitted = pyqtSignal(bool)
+
+    def __init__(self, iface, db, qgis_utils, wizard_settings):
+        super(WizardFactory, self).__init__(iface, db, qgis_utils, wizard_settings)
+
+    def init_gui(self):
+        raise NotImplementedError
+
+    def adjust_page_1_controls(self):
+        raise NotImplementedError
+
+    def finished_dialog(self):
+        self.save_settings()
+
+        if self.rad_refactor.isChecked():
+            if self.mMapLayerComboBox.currentLayer() is not None:
+                field_mapping = self.cbo_mapping.currentText()
+                res_etl_model = self.qgis_utils.show_etl_model(self._db,
+                                                               self.mMapLayerComboBox.currentLayer(),
+                                                               self.EDITING_LAYER_NAME,
+                                                               field_mapping=field_mapping)
+                if res_etl_model: # Features were added?
+                    # If the result of the etl_model is successful and we used a stored recent mapping, we delete the
+                    # previous mapping used (we give preference to the latest used mapping)
+                    if field_mapping:
+                        self.qgis_utils.delete_old_field_mapping(field_mapping)
+
+                    self.qgis_utils.save_field_mapping(self.EDITING_LAYER_NAME)
+            else:
+                self.qgis_utils.message_emitted.emit(
+                    QCoreApplication.translate(self.WIZARD_NAME,
+                                               "Select a source layer to set the field mapping to '{}'.").format(
+                        self.EDITING_LAYER_NAME),
+                    Qgis.Warning)
+
+        elif self.rad_create_manually.isChecked():
+            self.prepare_feature_creation()
+
+    def close_wizard(self, message=None, show_message=True):
+        if message is None:
+            message = QCoreApplication.translate(self.WIZARD_NAME, "'{}' tool has been closed.").format(self.WIZARD_TOOL_NAME)
+        if show_message:
+            self.qgis_utils.message_emitted.emit(message, Qgis.Info)
+
+        if hasattr(self, 'SELECTION_ON_MAP'):
+            self.init_map_tool()
+
+        self.disconnect_signals()
+        self.set_wizard_is_open_emitted.emit(False)
+        self.close()
+
+    def edit_feature(self):
+        self.iface.layerTreeView().setCurrentLayer(self._layers[self.EDITING_LAYER_NAME][LAYER])
+        self._layers[self.EDITING_LAYER_NAME][LAYER].committedFeaturesAdded.connect(self.finish_feature_creation)
+        self.open_form(self._layers[self.EDITING_LAYER_NAME][LAYER])
+
+    def advance_save(self, features):
+        raise NotImplementedError
+
+    def open_form(self, layer):
+        if not layer.isEditable():
+            layer.startEditing()
+
+        self.exec_form(layer)
+
+    def get_feature_exec_form(self, layer):
+        feature = self.qgis_utils.get_new_feature(layer)
+        return feature
+
+    def exec_form_advance(self, layer):
+        raise NotImplementedError
